Use of current user information instead of fixed user id for testing - part 2
- Remove "single()" methods on Supabase to prevent issues.
- Update of UserInterestEventService to use current user information

+++ lib/features/user/services/user_interest_event_service.dart
@@ -1,4 +1,4 @@
-// user_interest_event_service.dart
+// lib/features/user/services/user_interest_event_service.dart
 
 import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:sway/features/event/models/event_model.dart';
@@ -7,10 +7,22 @@ import 'package:sway/features/user/models/user_model.dart' as AppUser;
 import 'package:sway/features/user/services/user_service.dart';
 
 class UserInterestEventService {
-  final _supabase = Supabase.instance.client;
-  final int userId = 3;
+  final SupabaseClient _supabase = Supabase.instance.client;
+  final UserService _userService = UserService();
 
+  /// Récupère l'ID de l'utilisateur actuellement connecté
+  Future<int?> _getCurrentUserId() async {
+    final currentUser = await _userService.getCurrentUser();
+    return currentUser?.id;
+  }
+
+  /// Vérifie si l'utilisateur est intéressé par un événement spécifique
   Future<bool> isInterestedInEvent(int eventId) async {
+    final userId = await _getCurrentUserId();
+    if (userId == null) {
+      throw Exception('Utilisateur non authentifié.');
+    }
+
     final response = await _supabase
         .from('user_interest_event')
         .select()
@@ -21,39 +33,84 @@ class UserInterestEventService {
     return response.isNotEmpty;
   }
 
+  /// Ajoute un intérêt pour un événement
   Future<void> addInterest(int eventId) async {
-    await _supabase.from('user_interest_event').insert({
+    final userId = await _getCurrentUserId();
+    if (userId == null) {
+      throw Exception('Utilisateur non authentifié.');
+    }
+
+    final response = await _supabase.from('user_interest_event').insert({
       'user_id': userId,
       'event_id': eventId,
       'status': 'interested',
     });
+
+    if (response.error != null) {
+      throw Exception(
+          'Erreur lors de l\'ajout de l\'intérêt pour l\'événement: ${response.error!.message}');
+    }
   }
 
+  /// Supprime un intérêt pour un événement
   Future<void> removeInterest(int eventId) async {
-    await _supabase
+    final userId = await _getCurrentUserId();
+    if (userId == null) {
+      throw Exception('Utilisateur non authentifié.');
+    }
+
+    final response = await _supabase
         .from('user_interest_event')
         .delete()
         .eq('user_id', userId)
         .eq('event_id', eventId)
         .eq('status', 'interested');
+
+    if (response.error != null) {
+      throw Exception(
+          'Erreur lors de la suppression de l\'intérêt pour l\'événement: ${response.error!.message}');
+    }
   }
 
+  /// Marque un événement comme "going"
   Future<void> markEventAsGoing(int eventId) async {
-    await _supabase
+    final userId = await _getCurrentUserId();
+    if (userId == null) {
+      throw Exception('Utilisateur non authentifié.');
+    }
+
+    final response = await _supabase
         .from('user_interest_event')
         .update({'status': 'going'})
         .eq('user_id', userId)
         .eq('event_id', eventId);
+
+    if (response.error != null) {
+      throw Exception(
+          'Erreur lors du marquage de l\'événement comme "going": ${response.error!.message}');
+    }
   }
 
+  /// Marque un événement comme "attended"
   Future<void> markEventAsAttended(int eventId) async {
-    await _supabase
+    final userId = await _getCurrentUserId();
+    if (userId == null) {
+      throw Exception('Utilisateur non authentifié.');
+    }
+
+    final response = await _supabase
         .from('user_interest_event')
         .update({'status': 'attended'})
         .eq('user_id', userId)
         .eq('event_id', eventId);
+
+    if (response.error != null) {
+      throw Exception(
+          'Erreur lors du marquage de l\'événement comme "attended": ${response.error!.message}');
+    }
   }
 
+  /// Récupère le nombre d'intérêts pour un événement avec un statut spécifique
   Future<int> getEventInterestCount(int eventId, String status) async {
     final response = await _supabase
         .from('user_interest_event')
@@ -64,7 +121,13 @@ class UserInterestEventService {
     return response.length;
   }
 
+  /// Vérifie si l'utilisateur va à un événement spécifique
   Future<bool> isGoingToEvent(int eventId) async {
+    final userId = await _getCurrentUserId();
+    if (userId == null) {
+      throw Exception('Utilisateur non authentifié.');
+    }
+
     final response = await _supabase
         .from('user_interest_event')
         .select()
@@ -75,7 +138,13 @@ class UserInterestEventService {
     return response.isNotEmpty;
   }
 
+  /// Vérifie si l'utilisateur a assisté à un événement spécifique
   Future<bool> isAttendedEvent(int eventId) async {
+    final userId = await _getCurrentUserId();
+    if (userId == null) {
+      throw Exception('Utilisateur non authentifié.');
+    }
+
     final response = await _supabase
         .from('user_interest_event')
         .select()
@@ -86,6 +155,7 @@ class UserInterestEventService {
     return response.isNotEmpty;
   }
 
+  /// Récupère les événements intéressés par un utilisateur spécifique
   Future<List<Event>> getInterestedEventsByUserId(int userId) async {
     final response = await _supabase
         .from('user_interest_event')
@@ -100,10 +170,12 @@ class UserInterestEventService {
     final DateTime now = DateTime.now();
 
     return allEvents
-        .where((event) => eventIds.contains(event.id) && event.dateTime.isAfter(now))
+        .where((event) =>
+            eventIds.contains(event.id) && event.dateTime.isAfter(now))
         .toList();
   }
 
+  /// Récupère les événements que l'utilisateur est "going" à
   Future<List<Event>> getGoingEventsByUserId(int userId) async {
     final response = await _supabase
         .from('user_interest_event')
@@ -118,10 +190,12 @@ class UserInterestEventService {
     final DateTime now = DateTime.now();
 
     return allEvents
-        .where((event) => eventIds.contains(event.id) && event.dateTime.isBefore(now))
+        .where((event) =>
+            eventIds.contains(event.id) && event.dateTime.isBefore(now))
         .toList();
   }
 
+  /// Récupère les événements que l'utilisateur a "attended"
   Future<List<Event>> getAttendedEventsByUserId(int userId) async {
     final response = await _supabase
         .from('user_interest_event')
@@ -136,10 +210,12 @@ class UserInterestEventService {
     final DateTime now = DateTime.now();
 
     return allEvents
-        .where((event) => eventIds.contains(event.id) && event.dateTime.isBefore(now))
+        .where((event) =>
+            eventIds.contains(event.id) && event.dateTime.isBefore(now))
         .toList();
   }
 
+  /// Récupère les utilisateurs intéressés par un événement spécifique
   Future<List<AppUser.User>> getInterestedUsersForEvent(int eventId) async {
     final response = await _supabase
         .from('user_interest_event')
@@ -150,9 +226,10 @@ class UserInterestEventService {
     final List<int> userIds =
         response.map((item) => item['user_id'] as int).toList();
 
-    return await UserService().getUsersByIds(userIds);
+    return await _userService.getUsersByIds(userIds);
   }
 
+  /// Récupère les utilisateurs qui vont à un événement spécifique
   Future<List<AppUser.User>> getGoingUsersForEvent(int eventId) async {
     final response = await _supabase
         .from('user_interest_event')
@@ -163,6 +240,6 @@ class UserInterestEventService {
     final List<int> userIds =
         response.map((item) => item['user_id'] as int).toList();
 
-    return await UserService().getUsersByIds(userIds);
+    return await _userService.getUsersByIds(userIds);
   }
 }
